translator-agent/pyproject.toml
[project]
name = "translator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.7.2",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.1.1",
]
